change append behavior of signalRemovePLI
Pre edit:
Line 72 in spikeDetection.m calls `signal = readCSC(channelFiles{j}, runRemovePLI, true, runCAR);`
The 3rd var "true" is clearRemovePLI - this is hardcoded so default behavior before edit was to add an empty signalRemovePLI vector to every unpacked micro mat file, regardless of whether removePLI was run. Even if removePLI is run, it is saved to the mat file and then immediately cleared.

Post edit:
This edit retains the possibility of saving signalRemovePLI in micro mat files (if clearRemovePLI is set to false in spikeDetection), but is still hardcoded to not.

Now if removePLI=1 (and this has NOT been run before) and clearRemovePLI=1, signalRemovePLI is just not saved to the mat file.

If removePLI=1 (and this HAS been run before) and clearRemovePLI=1, signalRemovePLI is loaded from the mat and then set to an empty vector.

If removePLI=0 no empty signalRemovePLI is appended to the micro mat file.